implement get_transactions_throughput (#3683)
* feat(grpc): implement get_transactions_throughput 

* feat(grpc): file rename

--------
