+++ massa-node/src/main.rs
@@ -22,8 +22,8 @@ use massa_execution_worker::start_execution_worker;
 use massa_factory_exports::{FactoryChannels, FactoryConfig, FactoryManager};
 use massa_factory_worker::start_factory;
 use massa_final_state::{FinalState, FinalStateConfig};
-use massa_grpc::api::MassaGrpcService;
 use massa_grpc::config::GrpcConfig;
+use massa_grpc::handler::MassaGrpcService;
 use massa_ledger_exports::LedgerConfig;
 use massa_ledger_worker::FinalLedger;
 use massa_logging::massa_trace;
@@ -95,7 +95,7 @@ async fn launch(
     StopHandle,
     StopHandle,
     StopHandle,
-    Option<massa_grpc::api::StopHandle>,
+    Option<massa_grpc::handler::StopHandle>,
 ) {
     info!("Node version : {}", *VERSION);
     if let Some(end) = *END_TIMESTAMP {
@@ -735,7 +735,7 @@ async fn stop(
     api_private_handle: StopHandle,
     api_public_handle: StopHandle,
     api_handle: StopHandle,
-    grpc_handle: Option<massa_grpc::api::StopHandle>,
+    grpc_handle: Option<massa_grpc::handler::StopHandle>,
 ) {
     // stop bootstrap
     if let Some(bootstrap_manager) = bootstrap_manager {
