+++ massa-proto/proto/massa/api.proto
@@ -39,10 +39,22 @@ service Grpc {
         option (google.api.http) = {get: "/v1/version"};
     }
 
+    // GetTransactionsThroughput
+    rpc GetTransactionsThroughput (GetTransactionsThroughputRequest) returns (GetTransactionsThroughputResponse) {
+        option (google.api.http) = {get: "/v1/tx"};
+    }
+
     // GetNextBlockBestParents
     rpc GetNextBlockBestParents (GetNextBlockBestParentsRequest) returns (GetNextBlockBestParentsResponse) {
         option (google.api.http) = {get: "/v1/get_next_block_best_parents"};
     }
+
+
+    // ███████╗████████╗██████╗ ███████╗ █████╗ ███╗   ███╗
+    // ██╔════╝╚══██╔══╝██╔══██╗██╔════╝██╔══██╗████╗ ████║
+    // ███████╗   ██║   ██████╔╝█████╗  ███████║██╔████╔██║
+    // ╚════██║   ██║   ██╔══██╗██╔══╝  ██╔══██║██║╚██╔╝██║
+    // ███████║   ██║   ██║  ██║███████╗██║  ██║██║ ╚═╝ ██║
 
     // SendBlocks
     rpc SendBlocks (stream SendBlocksRequest) returns (stream SendBlocksResponse) {}
@@ -52,6 +64,29 @@ service Grpc {
 
     // SendOperations
     rpc SendOperations (stream SendOperationsRequest) returns (stream SendOperationsResponse) {}
+
+    // Subscribe GetTransactionsThroughput
+    rpc SubscribeTransactionsThroughput (stream GetTransactionsThroughputStreamRequest) returns (stream GetTransactionsThroughputResponse) {}
+}
+
+message GetTransactionsThroughputRequest {
+    //id
+    string id = 1;
+}
+
+message GetTransactionsThroughputStreamRequest {
+    //id
+    string id = 1;
+
+    // optional timer interval in sec
+    optional uint64 interval = 2;
+}
+
+message GetTransactionsThroughputResponse {
+    //id
+    string id = 1;
+    // tx/s
+    uint32 tx_s = 2;
 }
 
 message GetNextBlockBestParentsRequest {
